requirements.txt
# Core frameworks
autogen>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
llamaindex>=0.9.0

# LLM integrations
openai>=1.0.0
anthropic>=0.5.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0

# Web and API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database and persistence
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6
redis>=4.5.0
chromadb>=0.4.0

# Tool integration
requests>=2.30.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
playwright>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.1.0
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
